requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
python-dotenv==1.2.1
pytesseract==0.3.10
numpy<2.0.0
opencv-python==4.9.0.80
openai==1.30.1
